Pondering support
First stab at supporting asynchronous / abort-able search. This is a work
in progress. Working towards pondering / time control.

+++ Chess/Search/Algorithm.hs
@@ -1,29 +1,31 @@
+{-# LANGUAGE MultiWayIf #-}
+{-# LANGUAGE DoAndIfThenElse #-}
 module Chess.Search.Algorithm
        ( search
        ) where
+
 import           Chess.Board
+import qualified Chess.Board as B (hash)
 import           Chess.Evaluation
-import qualified Chess.Killer as K
-
+import qualified Chess.Killer              as K
 import           Chess.Move
-import qualified Chess.PVStore as PVS
+import qualified Chess.Move                as M (calcHash)
+import qualified Chess.PVStore             as PVS
 import           Chess.Search.Search
-import qualified Chess.Search.SearchResult as SR ((<@>), (<++>), moves)
-import           Chess.Search.SearchResult hiding (moves, (<@>), (<++>))
+import qualified Chess.Search.SearchResult as SR (moves)
+import           Chess.Search.SearchResult hiding (moves)
 import           Chess.Search.SearchState
-import           Chess.TransPosCache (typ, result)
-import qualified Chess.TransPosCache as TPC
-import           Control.Lens
-import           Control.Monad
-import           Control.Monad.State
+import qualified Chess.Search.SearchState  as SS (hash)
+import qualified Chess.TransPosCache       as TPC
+import           Control.Applicative       ((<$>))
+import           Control.Concurrent.STM    (readTVarIO)
+import           Control.Lens              ((+=), (.=), (^.), (%=), makeLenses, use, view)
+import           Control.Monad             (liftM, when)
+import           Control.Monad.State       (get, liftIO)
 import           Data.ChessTypes
+import           Data.Foldable             (forM_)
 import           Data.List.Extras
-
-
-(<@>) :: (Int -> Int) -> Search SearchResult -> Search SearchResult
-f <@> m = liftM (f SR.<@>) m
-(<++>) :: Move -> Search SearchResult -> Search SearchResult
-x <++> m = liftM (x SR.<++>) m
+import           Data.Maybe                (fromJust, isJust)
 
 
 data LoopResult = LoopResult
@@ -34,34 +36,58 @@ data LoopResult = LoopResult
 $(makeLenses ''LoopResult)
 
 
-
 -- | top level search
 search
   :: Int -- ^ depth
-  -> Search SearchResult
+  -> Search (Maybe SearchResult)
 search d = do
   tpcHit   .= 0
   tpcMiss  .= 0
   nCnt     .= 0
-  c <- liftM (\b -> direction (b^.next) 1) $ use board
-  withIterativeDeepening d 1 $ \d' -> do
-    r <- (c*) <@> negaScout d' d' (-inf) inf c
-    pv  .= PVS.insert (r^.SR.moves)
-    return r
+--  tpc      %= TPC.transPosCacheDeflate
+  mPonderHit <- use previous
+  b          <- use board
+  s <- if maybe False (\ph -> b^.B.hash == ph^.SS.hash) mPonderHit
+       then do
+         let ph = fromJust mPonderHit
+         report $ "ponderhit @ " ++ (show $ ph^.depth) ++ " " ++ (renderVariation (ph^.result))
+         pv  .= PVS.insert (ph^.result^.SR.moves)
+         return $ (ph^.depth)
+       else do
+         report $ "pondermiss"
+         return 1
+
+  let c = direction (b^.next) 1
+  r <- withIterativeDeepening d s $ \d' -> do
+       r' <- (c*) <@> negaScout d' d' (-inf) inf c
+       forM_ r' $ \r'' -> do
+         pv .= PVS.insert (r''^.SR.moves)
+         case M.calcHash b <$> (first r'') of
+           Just h -> do
+                   previous .= (Just $ Previous
+                                { _depth = d' - 2
+                                , _result = shift r''
+                                , _hash = h
+                                })
+           Nothing -> return ()
+       return r'
+  return r
 
 
 -- | searches with iterative deepening
 withIterativeDeepening
-  :: Int                -- ^ max depth
-  -> Int                -- ^ start depth
-  -> (Int -> Search a)  -- ^ search of given depth
-  -> Search a
+  :: Int                        -- ^ max depth
+  -> Int                        -- ^ start depth
+  -> (Int -> Search (Maybe a))  -- ^ search of given depth
+  -> Search (Maybe a)
 withIterativeDeepening mx d s = do
-  liftIO $ putStrLn $ "info depth " ++ show d
+  report $ "depth " ++ show d
   r <- s d
-  if d >= mx
-    then return r
-    else withIterativeDeepening mx (d + 1) s
+  maybe (return Nothing)
+        (const $ if d >= mx
+                 then return r
+                 else withIterativeDeepening mx (d + 1) s)
+        r
 
 
 -- | executes search action with a moved played
@@ -80,8 +106,8 @@ withTransPosCache
   :: Int -- ^ depth
   -> Int -- ^ alpha
   -> Int -- ^ beta
-  -> (Int -> Int -> Maybe Move -> Search SearchResult)
-  -> Search SearchResult
+  -> (Int -> Int -> Maybe Move -> Search (Maybe SearchResult))
+  -> Search (Maybe SearchResult)
 withTransPosCache d alpha beta f = do
   b <- use board
   t <- use tpc
@@ -91,16 +117,19 @@ withTransPosCache d alpha beta f = do
     Right (cache', entry) -> do
       tpc .= cache'
       tpcHit += 1
-      case entry^.typ of
-        TPC.Exact -> return $ entry^.result
-        TPC.Lower -> let alpha' = max alpha $ entry^.result^.eval
+      case entry^.TPC.typ of
+
+        TPC.Exact -> return $ Just $ entry^.TPC.result
+
+        TPC.Lower -> let alpha' = max alpha $ entry^.TPC.result^.eval
                      in if alpha' >= beta
-                        then return $ SearchResult [] alpha'
-                        else f alpha' beta $ first $ entry^.result
-        TPC.Upper -> let beta' = min beta $ entry^.result^.eval
+                        then return $ Just $ SearchResult [] alpha'
+                        else f alpha' beta $ first $ entry^.TPC.result
+
+        TPC.Upper -> let beta' = min beta $ entry^.TPC.result^.eval
                      in if alpha >= beta'
-                        then return $ SearchResult [] alpha
-                        else f alpha beta' $ first $ entry^.result
+                        then return $ Just $ SearchResult [] alpha
+                        else f alpha beta' $ first $ entry^.TPC.result
     Left mr -> do
       tpcMiss += 1
       f alpha beta mr
@@ -112,14 +141,16 @@ withKiller
   :: [ Move ] -- ^ move list
   -> Int      -- ^ max depth
   -> Int      -- ^ depth
-  -> ( [ Move ] -> Search LoopResult)
-  -> Search LoopResult  
+  -> ( [ Move ] -> Search (Maybe LoopResult))
+  -> Search (Maybe LoopResult)
 withKiller ml mx d f = do
   ml' <- liftM (K.heuristics (mx - d) ml) (use kill)
-  lr@(LoopResult t res) <- f ml'
-  when (t == TPC.Lower) $ kill %= K.insert (mx - d) (first res)
-  kill %= K.clearLevel (mx - d + 1)
-  return lr
+  mlr <- f ml'
+  when (isJust mlr) $ do
+    let Just (LoopResult t res) = mlr
+    when (t == TPC.Lower) $ kill %= K.insert (mx - d) (first res)
+    kill %= K.clearLevel (mx - d + 1)
+  return mlr
 {-# INLINE withKiller #-}
 
 
@@ -128,8 +159,8 @@ withPVStore
   :: [ Move ] -- ^ move list
   -> Int      -- ^ max depth
   -> Int      -- ^ depth
-  -> ( [Move] -> Search LoopResult )
-  -> Search LoopResult
+  -> ( [Move] -> Search (Maybe LoopResult ))
+  -> Search (Maybe LoopResult)
 withPVStore ml mx d f = do
   ml' <- liftM (PVS.heuristics (mx - d) ml) (use pv)
   r <- f ml'
@@ -143,8 +174,8 @@ withPVStore ml mx d f = do
 withMaybeMove
   :: [ Move ]   -- ^ move list
   -> Maybe Move -- ^ a good move if any ( from the tt )
-  -> ([Move] -> Search LoopResult)
-  -> Search LoopResult
+  -> ([Move] -> Search (Maybe LoopResult))
+  -> Search (Maybe LoopResult)
 withMaybeMove ml (Just m) f = f (toFront m ml)
 withMaybeMove ml Nothing  f = f ml
 {-# INLINE withMaybeMove #-}
@@ -156,8 +187,8 @@ withStores
   -> Maybe Move -- ^ a good move if any ( from the tt )
   -> Int        -- ^ max depth
   -> Int        -- ^ depth
-  -> ( [Move] -> Search LoopResult )
-  -> Search LoopResult
+  -> ( [Move] -> Search (Maybe LoopResult) )
+  -> Search (Maybe LoopResult)
 withStores ml mm mx d f = withKiller ml mx d $ \ml' -> withPVStore ml' mx d $ \ml'' -> withMaybeMove ml'' mm f
 {-# INLINE withStores #-}
 
@@ -169,26 +200,40 @@ iterateMoves
   -> Int                                    -- ^ alpha
   -> Int                                    -- ^ beta
   -> Bool                                   -- ^ report info
-  -> (Bool -> Move -> Int -> Int -> Search SearchResult) -- ^ True is fed in in the first iteration (for negascout)
-  -> Search LoopResult
+  -> (Bool -> Move -> Int -> Int -> Search (Maybe SearchResult)) -- ^ True is fed in in the first iteration (for negascout)
+  -> Search (Maybe LoopResult)
 iterateMoves ml d alpha beta rep ac = do
-  lr <- go True (LoopResult TPC.Upper $ SearchResult [] alpha) ml
-  b  <- use board
-  tpc %= TPC.transPosCacheInsert b d (lr^.tpcEntryT) (lr^.line)
-  return lr
-  where go _ l [] = return l
+  mlr <- go True (LoopResult TPC.Upper $ SearchResult [] alpha) ml
+  case mlr of
+    Just lr -> do
+        b  <- use board
+        tpc %= TPC.transPosCacheInsert b d (lr^.tpcEntryT) (lr^.line)
+    Nothing -> return ()
+  return $ mlr
+
+  where go _ l [] = return $ Just l
         go f prev (m:ms) = do
-          let prevVal = prev^.line^.eval
-          nxt <- ac f m prevVal beta
-          let cont
-                | nxt^.eval >= beta   = return $ LoopResult TPC.Lower $ SearchResult [m] beta
-                | nxt^.eval > prevVal = info nxt m >> go False (LoopResult TPC.Exact nxt) ms
-                | otherwise           = info (prev^.line) m >> go False prev ms
-          cont
-        info sr m = when rep $ get >>= \st -> liftIO $ putStrLn ("info TPC : " ++ show (hitRatio st) ++ "% "
-                                                                 ++ show (st^.nCnt `div` 1000) ++ "kn "
-                                                                 ++ " PV : " ++ renderVariation sr
-                                                                 ++ " curr : " ++ renderShortMove m)
+          abortedtv <- use aborted
+          abort     <- liftIO $ readTVarIO abortedtv
+          if abort
+          then return Nothing
+          else do
+            let prevVal = prev^.line^.eval
+            mnxt <- ac f m prevVal beta
+            maybe (return Nothing)
+                  (\nxt -> if
+                       | nxt^.eval >= beta   -> return $ Just $ LoopResult TPC.Lower $ SearchResult [m] beta
+                       | nxt^.eval > prevVal -> info nxt m >> go False (LoopResult TPC.Exact nxt) ms
+                       | otherwise           -> info (prev^.line) m >> go False prev ms)
+                  mnxt
+
+        info sr m = when rep $ do
+                      st <- get
+                      let s =  "TPC : " ++ show (hitRatio st) ++ "% "
+                               ++ show (st^.nCnt `div` 1000) ++ "kn "
+                               ++ " PV : " ++ renderVariation sr
+                               ++ " curr : " ++ renderShortMove m
+                      report $ s
 
 percentage :: Int -> Int -> Int
 percentage a b = if a == 0 then 0 else 100 * b `div` a
@@ -205,30 +250,34 @@ negaScout
   -> Int -- ^ alpha
   -> Int -- ^ beta
   -> Int -- ^ colour
-  -> Search SearchResult
+  -> Search (Maybe SearchResult)
 negaScout mx d alpha' beta' c = withTransPosCache d alpha' beta' $ \alpha beta mr -> do
   mate <- liftM (not . anyMove) $ use board
-  if mate
-    then do
+  if
+    | mate -> do
       nCnt += 1
-      liftM (\b -> SearchResult [] $ c * evaluate b) $ use board
-    else if d == 0
-          then quiscene mx d alpha beta c
-          else do
-            ml <- liftM moves $ use board
+      liftM (\b -> Just $ SearchResult [] $ c * evaluate b) $ use board
 
-            r <- withStores ml mr mx d
-                 $ \ml' -> iterateMoves ml' d alpha beta (mx == d)
-                           $ \f m a b -> withMove m
-                                         $ m <++>
-                                         if not f
-                                         then do
-                                           n <- ((-1)*) <@> negaScout mx (d - 1) (-a - 1) (-a) (-c)
-                                           if a < n^.eval && n^.eval < b
-                                             then ((-1)*) <@> negaScout mx (d - 1) (-b) (-a) (-c)
-                                             else return n
-                                         else ((-1)*) <@> negaScout mx (d - 1) (-b) (-a) (-c)
-            return $ r^.line
+    | d == 0 -> quiscene mx d alpha beta c
+
+    | otherwise -> do
+       ml <- liftM moves $ use board
+
+       mr' <- withStores ml mr mx d
+              $ \ml' -> iterateMoves ml' d alpha beta (mx == d)
+                        $ \f m a b -> withMove m
+                                      $ m <++>
+                                        if not f
+                                        then do
+                                          mn <- ((-1)*) <@> negaScout mx (d - 1) (-a - 1) (-a) (-c)
+                                          maybe (return Nothing)
+                                                (\n -> if a < n^.eval && n^.eval < b
+                                                       then ((-1)*) <@> negaScout mx (d - 1) (-b) (-a) (-c)
+                                                       else return mn)
+                                                mn
+
+                                        else ((-1)*) <@> negaScout mx (d - 1) (-b) (-a) (-c)
+       return $ (view line) <$> mr'
 
 
 -- | quiscene search
@@ -238,7 +287,7 @@ quiscene
   -> Int -- ^ alpha
   -> Int -- ^ beta
   -> Int -- ^ colour
-  -> Search SearchResult
+  -> Search (Maybe SearchResult)
 quiscene mx d alpha' beta' c = withTransPosCache 0 alpha' beta' $ \alpha beta mr -> do
   standPat <- liftM ((c*) . evaluate) (use board)
   nCnt += 1
@@ -246,15 +295,15 @@ quiscene mx d alpha' beta' c = withTransPosCache 0 alpha' beta' $ \alpha beta mr
     then do
        b  <- use board
        tpc %= TPC.transPosCacheInsert b 0 TPC.Lower (SearchResult [] beta)
-       return $ SearchResult [] beta
+       return $ Just $ SearchResult [] beta
     else do
        ml <- liftM forcingMoves $ use board
 
        let alpha'' = max alpha standPat
-       r <- withStores ml mr mx d
-            $ \ml' ->  iterateMoves ml' 0 alpha'' beta False
-                       $ \ _ m a b -> withMove m $ m <++> (((-1)*) <@> quiscene mx (d - 1) (-b) (-a) (-c))
-       return $ r^.line
+       mr' <- withStores ml mr mx d
+              $ \ml' ->  iterateMoves ml' 0 alpha'' beta False
+                         $ \ _ m a b -> withMove m $ m <++> (((-1)*) <@> quiscene mx (d - 1) (-b) (-a) (-c))
+       return $ (view line) <$> mr'
 
 
 inf :: Int
